third_party/WebKit/LayoutTests/http/tests/performance-timing/paint-timing/basetest.html

<!DOCTYPE html>
<head>
<title>Performance Paint Timing Test</title>
</head>
<body>
    <script src="../../resources/testharness.js"></script>
    <script src="../../resources/testharnessreport.js"></script>
    <div id="main"></div>

    <script>
      async_test(function(t) {
        testRunner.capturePixelsAsyncThen(t.step_func(function() {
            const bufferedEntries = performance.getEntriesByType('paint');
            assert_equals(bufferedEntries.length, 0, "No paint entries yet");

            const div = document.createElement("div");
            div.style.width = "100px";
            div.style.height = "100px";
            div.style.backgroundColor = "red";
            div.style.color = "blue";
            div.innerHTML = "test"
            document.getElementById("main").appendChild(div);

            testRunner.capturePixelsAsyncThen(t.step_func_done(function() {
                const bufferedEntries = performance.getEntriesByType('paint');
                assert_equals(bufferedEntries.length, 2, "FP and FCP.");
                assert_equals(bufferedEntries[0].entryType, "paint");
                assert_equals(bufferedEntries[0].name, "first-paint");
                assert_equals(bufferedEntries[1].entryType, "paint");
                assert_equals(bufferedEntries[1].name, "first-contentful-paint");
            }));
        }));
      }, "Basic test to check existence of FP and FCP.");
    </script>
</body>
</html>
